+++ liblogging/src/lib.rs
@@ -6,12 +6,12 @@ use std::str::FromStr;
 use log::LevelFilter;
 
 use crate::console_logger::ConsoleLogger;
-use crate::postgres_logger::PostgresLogger;
+use crate::datastore_logger::DatastoreLogger;
 use crate::syslog_logger::SyslogLogger;
 
 mod console_logger;
 mod datastore;
-mod postgres_logger;
+mod datastore_logger;
 mod syslog_logger;
 
 static DEFAULT_ALLOWED_TARGETS: once_cell::sync::Lazy<Vec<String>> =
@@ -22,11 +22,11 @@ static DEFAULT_ALLOWED_TARGETS: once_cell::sync::Lazy<Vec<String>> =
             .collect()
     });
 
-/// Logger implementation that logs to console, syslog, and `PostgreSQL`
+/// Logger implementation that logs to console, syslog, and Datastore
 pub struct Logger {
     console: ConsoleLogger,
     syslog: SyslogLogger,
-    postgres: PostgresLogger,
+    datastore: DatastoreLogger,
     allowed_targets: Vec<String>,
 }
 
@@ -39,7 +39,7 @@ impl Logger {
         let LoggerConfig {
             console,
             syslog,
-            postgres,
+            datastore,
             allowed_targets,
         } = logger_config;
 
@@ -50,7 +50,7 @@ impl Logger {
             log::set_boxed_logger(Box::new(Logger {
                 console: ConsoleLogger::new(console),
                 syslog: SyslogLogger::new(syslog),
-                postgres: PostgresLogger::new(postgres),
+                datastore: DatastoreLogger::new(datastore),
                 allowed_targets,
             }))
             .unwrap_or_default();
@@ -63,7 +63,7 @@ impl log::Log for Logger {
     fn enabled(&self, metadata: &log::Metadata) -> bool {
         self.syslog.enabled(metadata)
             || self.console.enabled(metadata)
-            || self.postgres.enabled(metadata)
+            || self.datastore.enabled(metadata)
     }
 
     fn log(&self, record: &log::Record) {
@@ -75,14 +75,14 @@ impl log::Log for Logger {
         {
             self.syslog.log(record);
             self.console.log(record);
-            self.postgres.log(record);
+            self.datastore.log(record);
         }
     }
 
     fn flush(&self) {
         self.syslog.flush();
         self.console.flush();
-        self.postgres.flush();
+        self.datastore.flush();
     }
 }
 
@@ -92,8 +92,8 @@ pub struct LoggerConfig {
     pub console: bool,
     /// Whether to log to syslog
     pub syslog: bool,
-    /// Whether to log to `PostgreSQL`
-    pub postgres: bool,
+    /// Whether to log to Datastore
+    pub datastore: bool,
     /// The list of allowed targets.<br>
     ///   By default, only logs from `nullnet*`, `appguard*`, and `wallguard*` will be emitted.<br>
     ///   Use this parameter to specify additional targets
@@ -105,9 +105,9 @@ impl LoggerConfig {
     /// Creates a new logger configuration
     ///
     /// # Arguments
-    /// * `console_logger` - Whether to log to console
-    /// * `syslog_endpoint` - Whether to log to syslog
-    /// * `postgres_endpoint` - Whether to log to `PostgreSQL`
+    /// * `console` - Whether to log to console
+    /// * `syslog` - Whether to log to syslog
+    /// * `datastore` - Whether to log to Datastore
     /// * `allowed_targets` - The list of allowed targets.<br>
     ///   By default, only logs from `nullnet*`, `appguard*`, and `wallguard*` will be emitted.<br>
     ///   Use this parameter to specify additional targets
@@ -116,13 +116,13 @@ impl LoggerConfig {
     pub fn new(
         console: bool,
         syslog: bool,
-        postgres: bool,
+        datastore: bool,
         allowed_targets: Vec<&'static str>,
     ) -> Self {
         Self {
             console,
             syslog,
-            postgres,
+            datastore,
             allowed_targets,
         }
     }
@@ -133,7 +133,7 @@ impl Default for LoggerConfig {
         Self {
             console: true,
             syslog: true,
-            postgres: true,
+            datastore: true,
             allowed_targets: vec![],
         }
     }