+++ liblogging/CHANGELOG.md
@@ -3,7 +3,7 @@
 All releases with the relative changes are documented in this file.
 
 ## [0.2.0] - UNRELEASED
-- Added PostgreSQL logger
+- Added Datastore logger
 - Made console logging optional
 
 ## [0.1.2] - 2025-02-19